refactor: colocate frontend domain logic
